Move some plugins to the notebook extension
