+++ packages/filebrowser-extension/package.json
@@ -36,7 +36,7 @@
     "watch": "tsc -b --watch"
   },
   "dependencies": {
-    "@jupyterlab-classic/application": "0.1.0",
+    "@jupyterlab-classic/application": "^0.1.0",
     "@jupyterlab/application": "^3.0.0-rc.13",
     "@jupyterlab/apputils": "^3.0.0-rc.13",
     "@jupyterlab/coreutils": "^5.0.0-rc.13",
