+++ packages/notebook-extension/package.json
@@ -1,7 +1,7 @@
 {
-  "name": "@jupyterlab-classic/application-extension",
+  "name": "@jupyterlab-classic/notebook-extension",
   "version": "0.1.0",
-  "description": "JupyterLab Classic - Application Extension",
+  "description": "JupyterLab Classic - Notebook Extension",
   "homepage": "https://github.com/jtpio/jupyterlab-classic",
   "bugs": {
     "url": "https://github.com/jtpio/jupyterlab-classic/issues"
@@ -37,15 +37,10 @@
   },
   "dependencies": {
     "@jupyterlab-classic/application": "^0.1.0",
-    "@jupyterlab-classic/ui-components": "^0.1.0",
     "@jupyterlab/application": "^3.0.0-rc.12",
     "@jupyterlab/apputils": "^3.0.0-rc.12",
-    "@jupyterlab/codeeditor": "^3.0.0-rc.12",
-    "@jupyterlab/codemirror": "^3.0.0-rc.12",
-    "@jupyterlab/docregistry": "^3.0.0-rc.12",
-    "@jupyterlab/mainmenu": "^3.0.0-rc.12",
-    "@jupyterlab/settingregistry": "^3.0.0-rc.12",
-    "@jupyterlab/translation": "^3.0.0-rc.12",
+    "@jupyterlab/docmanager": "^3.0.0-rc.12",
+    "@jupyterlab/notebook": "^3.0.0-rc.12",
     "@lumino/widgets": "^1.14.0"
   },
   "devDependencies": {
