+++ packages/notebook-extension/style/base.css
@@ -1,0 +1,79 @@
+/*-----------------------------------------------------------------------------
+| Copyright (c) Jupyter Development Team.
+|
+| Distributed under the terms of the Modified BSD License.
+|----------------------------------------------------------------------------*/
+
+.jp-ClassicKernelLogo {
+  flex: 0 0 auto;
+  display: flex;
+  align-items: center;
+  text-align: center;
+  margin-right: 8px;
+}
+
+.jp-ClassicKernelLogo img {
+  max-width: 28px;
+  max-height: 28px;
+  display: flex;
+}
+
+.jp-ClassicKernelStatus {
+  font-size: 12px;
+  margin: 0;
+  font-weight: normal;
+  color: var(--jp-ui-font-color0);
+  font-family: var(--jp-ui-font-family);
+  line-height: var(--jp-private-title-panel-height);
+  padding-left: 5px;
+  padding-right: 5px;
+}
+
+.jp-ClassicKernelStatus-error {
+  background-color: var(--jp-error-color0);
+}
+
+.jp-ClassicKernelStatus-warn {
+  background-color: var(--jp-warn-color0);
+}
+
+.jp-ClassicKernelStatus-info {
+  background-color: var(--jp-info-color0);
+}
+
+.jp-ClassicKernelStatus-fade {
+  animation: 0.5s fade-out forwards;
+}
+
+@keyframes fade-out {
+  0% {
+    opacity: 1;
+  }
+  100% {
+    opacity: 0;
+  }
+}
+
+#jp-title h1 {
+  cursor: pointer;
+  font-size: 18px;
+  margin: 0;
+  font-weight: normal;
+  color: var(--jp-ui-font-color0);
+  font-family: var(--jp-ui-font-family);
+  line-height: calc(1.5 * var(--jp-private-title-panel-height));
+}
+
+#jp-title h1:hover {
+  background: var(--jp-layout-color2);
+}
+
+.jp-ClassicCheckpoint {
+  font-size: 14px;
+  margin-left: 5px;
+  margin-right: 5px;
+  font-weight: normal;
+  color: var(--jp-ui-font-color0);
+  font-family: var(--jp-ui-font-family);
+  line-height: calc(1.5 * var(--jp-private-title-panel-height));
+}
