+++ packages/notebook-extension/style/index.css
@@ -1,0 +1,1 @@
+@import url('./base.css');
